Move C++ code into a "cc" subdirectory.
The top-level directory contains the main Go package. The go tool
complains about the presence of the .cc and .h files and thinks they
should be part of the build. Having the C++ sources in their own
subdirectory does the right thing; the Go tool ignores directories
that do not contain Go code.